Limit JVM target to the published variant (main source set)
This will compile tests to use the same target as the current
JVM, and more importantly will change test dependencies'
resolution to a variant compatible with the current JVM rather
than with the plugin's target. This will allow updating JUnit
Jupiter to version 6 which requires Java 17, and would fail to
resolve when targetting Java 8.